This makes the codes more cleaner
This is the codes illustrated how the climate variables were extracted in the published article. Effects of landscape simplicity on crop yields: A reanalysis of global dataset. Cite this codes as Madin, M. B., & Nelson, K. S. (2023). Effects of landscape simplicity on crop yield: A reanalysis of a global database. Plos one, 18(12), e0289799. Or access the publication at https://doi.org/10.1371/journal.pone.0289799